Optimizations in usecases in order not to have the same functions in >1 usecases
