+++ app/src/main/java/com/weatherxm/data/Modules.kt
@@ -528,7 +528,7 @@ private val usecases = module {
         UserUseCaseImpl(get(), get(), get(), get())
     }
     single<PreferencesUseCase> {
-        PreferencesUseCaseImpl(get(), get(), get())
+        PreferencesUseCaseImpl(get(), get())
     }
     single<DeleteAccountUseCase> {
         DeleteAccountUseCaseImpl(get(), get())
@@ -549,7 +549,7 @@ private val usecases = module {
         StationSettingsUseCaseImpl(get(), get(), get())
     }
     single<WidgetSelectStationUseCase> {
-        WidgetSelectStationUseCaseImpl(get(), get(), get())
+        WidgetSelectStationUseCaseImpl(get(), get())
     }
     single<WidgetCurrentWeatherUseCase> {
         WidgetCurrentWeatherUseCaseImpl(get(), get())
@@ -975,7 +975,7 @@ private val viewmodels = module {
     viewModel { LoginViewModel(get(), get(), get()) }
     viewModel { NetworkStatsViewModel(get()) }
     viewModel { PasswordPromptViewModel(get(), get(), get()) }
-    viewModel { PreferenceViewModel(get(), get()) }
+    viewModel { PreferenceViewModel(get(), get(), get()) }
     viewModel { ResetPasswordViewModel(get(), get(), get()) }
     viewModel { RewardsClaimViewModel(get(), get()) }
     viewModel { RewardsListViewModel(get(), get()) }
@@ -994,7 +994,7 @@ private val viewmodels = module {
     viewModel { SignupViewModel(get(), get(), get()) }
     viewModel { UpdatePromptViewModel(get()) }
     viewModel { DeepLinkRouterViewModel(get(), get(), get()) }
-    viewModel { SelectStationViewModel(get()) }
+    viewModel { SelectStationViewModel(get(), get()) }
     viewModel { ClaimLocationViewModel(get(), get(), get()) }
     viewModel { ClaimHeliumViewModel(get(), get(), get()) }
     viewModel { ClaimHeliumPairViewModel(get(), get(), get(), get()) }
