+++ app/src/main/java/com/weatherxm/ui/widgets/selectstation/SelectStationViewModel.kt
@@ -6,12 +6,16 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.weatherxm.ui.common.Resource
 import com.weatherxm.ui.common.UIDevice
+import com.weatherxm.usecases.AuthUseCase
 import com.weatherxm.usecases.WidgetSelectStationUseCase
 import com.weatherxm.util.Failure.getDefaultMessage
 import kotlinx.coroutines.launch
 import timber.log.Timber
 
-class SelectStationViewModel(private val usecase: WidgetSelectStationUseCase) : ViewModel() {
+class SelectStationViewModel(
+    private val usecase: WidgetSelectStationUseCase,
+    private val authUseCase: AuthUseCase,
+) : ViewModel() {
 
     private val onDevices = MutableLiveData<Resource<List<UIDevice>>>()
     private val isNotLoggedIn = MutableLiveData<Unit>()
@@ -29,7 +33,7 @@ class SelectStationViewModel(private val usecase: WidgetSelectStationUseCase) : 
 
     fun checkIfLoggedInAndProceed() {
         Timber.d("Checking if user is logged in in the background")
-        if (usecase.isLoggedIn()) {
+        if (authUseCase.isLoggedIn()) {
             fetch()
         } else {
             isNotLoggedIn.postValue(Unit)
