+++ app/src/test/java/com/weatherxm/ui/preferences/PreferencesViewModelTest.kt
@@ -2,6 +2,7 @@ package com.weatherxm.ui.preferences
 
 import com.weatherxm.analytics.AnalyticsWrapper
 import com.weatherxm.ui.InstantExecutorListener
+import com.weatherxm.usecases.AuthUseCase
 import com.weatherxm.usecases.PreferencesUseCase
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
@@ -21,6 +22,7 @@ import kotlinx.coroutines.test.setMain
 @OptIn(ExperimentalCoroutinesApi::class)
 class PreferencesViewModelTest : BehaviorSpec({
     val usecase = mockk<PreferencesUseCase>()
+    val authUseCase = mockk<AuthUseCase>()
     val analytics = mockk<AnalyticsWrapper>()
     lateinit var viewModel: PreferenceViewModel
 
@@ -34,11 +36,11 @@ class PreferencesViewModelTest : BehaviorSpec({
         justRun { analytics.setAnalyticsEnabled(any()) }
         justRun { analytics.onLogout() }
         justRun { usecase.setAnalyticsEnabled(any()) }
-        coJustRun { usecase.logout() }
-        every { usecase.isLoggedIn() } returns true
+        coJustRun { authUseCase.logout() }
+        every { authUseCase.isLoggedIn() } returns true
         every { usecase.getInstallationId() } returns installationId
 
-        viewModel = PreferenceViewModel(usecase, analytics)
+        viewModel = PreferenceViewModel(usecase, authUseCase, analytics)
     }
 
     context("Invoke a change in SharedPreferences and update user's properties in analytics") {
@@ -68,7 +70,7 @@ class PreferencesViewModelTest : BehaviorSpec({
                 verify(exactly = 1) { analytics.onLogout() }
             }
             then("call the logout function in the usecase") {
-                coVerify(exactly = 1) { usecase.logout() }
+                coVerify(exactly = 1) { authUseCase.logout() }
             }
             then("LiveData onLogout gets invoked with `true` param") {
                 viewModel.onLogout().value shouldBe true
